fix: DRY up the requests fetching logic between subs/codes
