+++ src/components/SubsidyRequestManagementTable/index.js
@@ -1,0 +1,12 @@
+import SubsidyRequestManagementTable from './SubsidyRequestManagementTable';
+
+export {
+  transformRequestOverview,
+  transformRequests,
+} from './data/utils';
+
+export {
+  useSubsidyRequests,
+} from './data/hooks';
+
+export default SubsidyRequestManagementTable;
