+++ src/hooks/index.js
@@ -1,2 +1,3 @@
 export { default as useInterval } from './useInterval';
 export { default as useWindowSize } from './useWindowSize';
+export { default as useIsMounted } from './useIsMounted';
