+++ src/hooks/useIsMounted.jsx
@@ -1,0 +1,11 @@
+import { useEffect, useState } from 'react';
+
+const useIsMounted = () => {
+  const [isMounted, setIsMounted] = useState(false);
+  useEffect(() => {
+    setIsMounted(true);
+  }, []);
+  return isMounted;
+};
+
+export default useIsMounted;
